enhance: Add mutatable factory class

+++ SharpeAi/src/main/java/org/neil/util/mutator/Mutator.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 
 public interface Mutator<T,I> extends Serializable {
 
-    T createMutatedCopy(float mutationRate, T original, I input);
+    T createMutatedCopy(double mutationRate, T original, I input);
 
     T create(I input);
